Refactor utils into separate modules
