+++ tests/30_client/client_test_cases.py
@@ -1,5 +1,3 @@
-import uuid
-
 import pytest
 from httpx import URL
